+++ server-test/src/test/java/de/njsm/stocks/servertest/v2/FoodTest.java
@@ -44,12 +44,13 @@ public class FoodTest {
         String newName = "Cabal";
         int id = createNewFoodType(name);
 
-        assertOnRename(id, 0, newName)
+        assertOnRename(id, 0, newName, 42)
                 .statusCode(200)
                 .body("status", equalTo(0));
 
         assertOnFood()
-                .body("data.name", hasItem(newName));
+                .body("data.name", hasItem(newName))
+                .body("data.expirationOffset", hasItem(42));
     }
 
     @Test
@@ -72,7 +73,7 @@ public class FoodTest {
         String newName = "Cabal";
         int id = createNewFoodType(name);
 
-        assertOnRename(id, 99, newName)
+        assertOnRename(id, 99, newName, 0)
                 .statusCode(400)
                 .body("status", equalTo(3));
     }
@@ -81,7 +82,7 @@ public class FoodTest {
     public void renamingUnknownIdIsReported() {
         String newName = "Cabal";
 
-        assertOnRename(9999, 0, newName)
+        assertOnRename(9999, 0, newName, 0)
                 .statusCode(404)
                 .body("status", equalTo(2));
     }
@@ -152,13 +153,14 @@ public class FoodTest {
 
     }
 
-    private ValidatableResponse assertOnRename(int id, int version, String newName) {
+    private ValidatableResponse assertOnRename(int id, int version, String newName, int expirationOffset) {
         return
         given()
                 .log().ifValidationFails()
                 .queryParam("id", id)
                 .queryParam("version", version)
-                .queryParam("new", newName).
+                .queryParam("new", newName)
+                .queryParam("expirationoffset", expirationOffset).
         when()
                 .put(TestSuite.DOMAIN + "/v2/food/rename").
         then()